requirements.txt
nibabel==3.2.1
numpy>=1.21
pandas>=1.3.4
Pillow>=9.0.0
